Switch to JDK 11
JDK 8 still can be used to build the project but JDK 11 will be used
by default in the CI pipeline. The only exception is the integration
tests with SonarQube 6.7.x (LTS), they'll continue to run using JDK 8.

Note that this doesn't require you to run SonarQube using JDK 11. If you
build this project using JDK 11, it'll run correctly in a SonarQube
running on JDK 8, just as it always did.